Fixed occasional crash + other improvements
Upgraded to preview 5 of StereoKit to fix occasional crashes,

Nodes can now be edited with left or right hand.

Editing options are hidden away while you are actively editing another property. This stop accidental editing icon selections and makes it clearer which input field has the keyboard focus.

Changed locomotion on the left hand to make it more consistent with the right hand and allow the hand to be used for interacting without accidentally performing locomotion.